move erc721 receiver to the separate file
